refactor(files): consolidate domain logic imports
