+++ apps/backend/files-service/src/files/application/permission-propagation.service.ts
@@ -2,7 +2,7 @@ import { Injectable, Logger } from '@nestjs/common';
 import { FolderEntity, PermissionEntity } from '../domain/entities';
 import { PermissionAction } from '../domain/enums';
 import { FileRepository, FolderRepository } from '../domain/interfaces';
-import { calculateBulkUpdates } from '../domain/logic/access-control.logic';
+import { calculateBulkUpdates } from '../domain/logic';
 
 @Injectable()
 export class PermissionPropagationService {
